requirements.txt
numpy>=1.24
scipy>=1.10
flask>=3.0
flask-socketio>=5.3
pywebview>=5.0
pyinstaller>=6.0
